credit-intelligence-starter/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
yfinance==0.2.28
alpha-vantage==2.3.1
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
textblob==0.17.1
vaderSentiment==3.3.2
plotly==5.17.0
python-dotenv==1.0.0
asyncio==3.4.3
aiohttp==3.9.1
schedule==1.2.0
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
